Split up tests into categories with common helper functions.
In preparation for adding more different types of tests, split up the
existing tests into three main categories. The npm run scripts now
accept as an argument which test file you want to run.
