+++ tests/testcafe/publiclibrary.js
@@ -1,0 +1,88 @@
+import { Selector, ClientFunction } from 'testcafe';
+
+import { escapeID } from '../../client/js/domhelpers.js';
+import { compareState, prepareClient, setName, setupTestEnvironment } from './util.js';
+
+setupTestEnvironment();
+
+async function removeGame(t, index) {
+  await t
+    .pressKey('esc')
+    .click('#statesButton')
+    .hover(`.roomState:nth-of-type(${index || 1})`)
+    .click(`.roomState:nth-of-type(${index || 1}) .edit`)
+    .click('.remove-game');
+}
+
+function publicLibraryTest(game, variant, md5, tests) {
+  test.after(async t => {
+    await removeGame(t);
+    await t.expect(Selector('#statesOverlay').visible).ok();
+  })(`Public library: ${game} (variant ${variant})`, async t => {
+    await ClientFunction(prepareClient)();
+    await t
+      .pressKey('esc')
+      .click('#statesButton')
+      .click(Selector('td.name').withExactText(game).prevSibling().child())
+      .hover('.roomState')
+      .click(Selector('button.play').nth(variant));
+    await setName(t);
+    await tests(t);
+    await compareState(t, md5);
+  });
+}
+
+function publicLibraryButtons(game, variant, md5, tests) {
+  publicLibraryTest(game, variant, md5, async t => {
+      for(const b of tests)
+        if(typeof b == "string") {
+          if(b.charAt(0) == '#') {
+            await t.click(b);
+          } else {
+            await t.click(`#w_${escapeID(b)}`);
+          }
+        } else {
+          await t.dragToElement(b[0](), b[1](), { speed:0.5 });
+        }
+  });
+}
+
+publicLibraryButtons('Blue',               0, 'c0b6df4181e1e14db376c87cef71a1c3', [
+  'fcc3fa2c-c091-41bc-8737-54d8b9d3a929', 'd3ab9f5f-daa4-4d81-8004-50a9c90af88e_incrementButton',
+  'd3ab9f5f-daa4-4d81-8004-50a9c90af88e_incrementButton', 'd3ab9f5f-daa4-4d81-8004-50a9c90af88e_decrementButton',
+  'reset_button', '#buttonInputGo', 'fcc3fa2c-c091-41bc-8737-54d8b9d3a929', '9n2q'
+]);
+publicLibraryButtons('Bhukhar',            0, '2748e96293b8646894700440508dd280', [ 'btnMenuSettings', 'btn8Players', 'btn4Packs', 'btnCloseSettings', 'btnSelectPlayer', 'btnDeal', 'btnPile4', 'btnStartGame', 'btnTakeOne', 'btnNextPlayer', 'btnPickUp' ]);
+publicLibraryButtons('Dice',               0, 'a68d28c20b624d6ddf87149bae230598', [ 'k18u', 'hy65', 'gghr', 'dsfa', 'f34a', 'fusq' ]);
+publicLibraryButtons('Dots',               0, '23894df38f786cb014fa1cd79f2345db', [ 'reset', '#buttonInputGo', 'col11', 'col21', 'col12', 'col22', 'row11', 'row31', 'row21', 'row32', 'row12', 'row42', 'row22', 'row23', 'col23' ]);
+publicLibraryButtons('Solitaire',          0, 'd5babf02d0c94500673d31188405ad9a', [ 'reset', 'jemz', 'reset' ]);
+publicLibraryButtons('Mancala',            0, '92108a0e76fd295fee9881b6c7f8928b', ['btnRule1', 'btnRule2', 'getb5', 'getb5', 'getb5', 'getb5', 'getb1', 'getb1', 'getb1', 'getb1' ]);
+publicLibraryButtons('Reversi',            0, '35e0017570f9ecd206a2317c1528be36',
+       [
+         [ ()=>Selector("#w_zpiece15"), ()=>Selector("#w_sq23") ],
+         [ ()=>Selector("#w_zpiece78"), ()=>Selector("#w_sq22") ],
+         [ ()=>Selector("#w_zpiece40"), ()=>Selector("#w_sq32") ],
+         [ ()=>Selector("#w_zpiece72"), ()=>Selector("#w_sq12") ],
+         [ ()=>Selector("#w_zpiece72"), ()=>Selector("#w_sq24") ],
+         [ ()=>Selector("#w_zpiece19"), ()=>Selector("#w_sq35") ],
+         [ ()=>Selector("#w_zpiece08"), ()=>Selector("#w_sq53") ]
+       ]);
+publicLibraryButtons('Reward',             0, '60c8c548fec57ead23c697b43216544f', [
+  'gmex', 'kprc', 'oksq', 'j1wz', 'vfhn', 'seat1', 'next'
+]);
+publicLibraryButtons('Rummy Tiles',        0, '6c5f8896284f52e874076bc60cca6325', [ 'startMix', 'draw14' ]);
+publicLibraryButtons('Undercover',         1, '97725a1d0733ef74dd5e1d0f9f260cb5', [ 'Reset', 'Spy Master Button' ]);
+publicLibraryButtons('Functions - CALL',   0, '4ef11451bfae5b8708e8f0eac2a06df4', [
+  'n4cw_8_C', '5a52', '5a52', '66kr', 'qeg1', 'n4cwB', '8r6p', 'qeg1', 'qeg1', 'n5eu'
+]);
+publicLibraryButtons('Functions - CLICK',  0, 'd44e77e0782cadbc9594494e5a83dde0', [ '7u2q' ]);
+publicLibraryButtons('Functions - ROTATE', 0, '70782503b9e3fb2d4e24495f5c53ef1b', [ 'c44c', '9kdj', 'w53c', 'w53c' ]);
+publicLibraryButtons('Functions - SELECT', 2, '51f108279a644454323430b8d993068a', [ 'jkmt1']);
+publicLibraryButtons('Functions - SORT',   1, '433948b27014f8236bd1978bdbe39443', [
+  'ingw', 'k131', 'cnfu', 'i6yz', 'z394', '0v3h', '1h8o', 'v5ra', 'ingw-copy001', 'k131-copy001', 'cnfu-copy001',
+  'i6yz-copy001', 'z394-copy001', '0v3h-copy001'
+]);
+publicLibraryButtons('Master Button',      0, 'eb19dffdb38641d5556e5fdb2c47c62b', [
+  'masterbutton', 'redbutton', 'orangebutton', 'yellowbutton', 'greenbutton', 'bluebutton', 'indigobutton',
+  'violetbutton', 'fae4', 'vbx5'
+]);
